requirements.txt
aiofiles==0.3.1
appnope==0.1.0
asn1crypto==0.22.0
cffi==1.10.0
click==6.7
croniter==0.3.16
cryptography==1.8.1
decorator==4.0.11
httptools==0.0.9
idna==2.5
ipython==6.0.0
ipython-genutils==0.2.0
itsdangerous==0.24
japronto==0.1.1
jedi==0.10.2
Jinja2==2.9.6
MarkupSafe==1.0
nose==1.3.7
numpy==1.12.1
packaging==16.8
paramiko==2.1.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyasn1==0.2.3
pycparser==2.17
Pygments==2.2.0
pymongo==3.2
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
sanic==0.5.4
semantic-version==2.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
ujson==1.35
uvloop==0.8.0
wcwidth==0.1.7
websockets==3.3
gunicorn
